// t/front-end/nutrition-cart/src/models/meals.ts
export interface Meal {
  id: number;
  strMeal: string;
  strInstructions: string;
  img: string;
  strIngredient1: string;
  strIngredient2: string;
  strIngredient3: string;
  strIngredient4: string;
  strIngredient5: string;
  strIngredient6: string;
  strIngredient7: string;
  strIngredient8: string;
  strIngredient9: string;
  strIngredient10: string;
  strIngredient11: string;
  strIngredient12: string;
  strIngredient13: string;
  strIngredient14: string;
  strIngredient15: string;
  strIngredient16: string;
  strIngredient17: string;
  strIngredient18: string;
  strIngredient19: string;
  strIngredient20: string;
  intMeasure1: number;
  intMeasure2: number;
  intMeasure3: number;
  intMeasure4: number;
  intMeasure5: number;
  intMeasure6: number;
  intMeasure7: number;
  intMeasure8: number;
  intMeasure9: number;
  intMeasure10: number;
  intMeasure11: number;
  intMeasure12: number;
  intMeasure13: number;
  intMeasure14: number;
  intMeasure15: number;
  intMeasure16: number;
  intMeasure17: number;
  intMeasure18: number;
  intMeasure19: number;
  intMeasure20: number;
}